Trait elgamal_shuffle::utils::transcript::TranscriptProtocol

pub trait TranscriptProtocol {

    // Required methods
    fn shuffle_domain_sep(&mut self, n: u64, m: u64);
    fn mexp_domain_sep(&mut self, m: u64, mu: u64);
    fn prod_domain_sep(&mut self, n: u64, m: u64);
    fn append_scalar(&mut self, label: &'static [u8], scalar: &Scalar);
    fn append_scalar_vec(&mut self, label: &'static [u8], scalar: &Vec<Scalar>);
    fn append_point(
        &mut self,
        label: &'static [u8],
        point: &CompressedRistretto,
    );
    fn append_point_vec(
        &mut self,
        label: &'static [u8],
        points: &Vec<CompressedRistretto>,
    );
    fn append_ciphertext(&mut self, label: &'static [u8], text: &Ciphertext);
    fn append_cipher_vec(
        &mut self,
        label: &'static [u8],
        texts: &Vec<Ciphertext>,
    );
    fn val_append_cipher_vec(
        &mut self,
        label: &'static [u8],
        texts: &Vec<Ciphertext>,
    ) -> Result<(), ProofError>;
    fn val_append_point_vec(
        &mut self,
        label: &'static [u8],
        points: &Vec<CompressedRistretto>,
    ) -> Result<(), ProofError>;
    fn validate_and_append_point(
        &mut self,
        label: &'static [u8],
        point: &CompressedRistretto,
    ) -> Result<(), ProofError>;
    fn validate_and_append_cipher(
        &mut self,
        label: &'static [u8],
        point: &Ciphertext,
    ) -> Result<(), ProofError>;
    fn challenge_scalar(&mut self, label: &'static [u8]) -> Scalar;

}

Required Methods

fn shuffle_domain_sep(&mut self, n: u64, m: u64)

Append a domain seperator for an m*n = N card, shuffle proof

fn mexp_domain_sep(&mut self, m: u64, mu: u64)

Append a domain seperator for an m = mu * m’ reduction, Multi-expo proof

fn prod_domain_sep(&mut self, n: u64, m: u64)

Append a domain seperator for an nxm sized, product proof

fn append_scalar(&mut self, label: &'static [u8], scalar: &Scalar)

Append a scalar with the given label.

fn append_scalar_vec(&mut self, label: &'static [u8], scalar: &Vec<Scalar>)

fn append_point(&mut self, label: &'static [u8], point: &CompressedRistretto)

Append a point with the given label.

fn append_point_vec( &mut self, label: &'static [u8], points: &Vec<CompressedRistretto>, )

fn append_ciphertext(&mut self, label: &'static [u8], text: &Ciphertext)

fn append_cipher_vec(&mut self, label: &'static [u8], texts: &Vec<Ciphertext>)

fn val_append_cipher_vec( &mut self, label: &'static [u8], texts: &Vec<Ciphertext>, ) -> Result<(), ProofError>

fn val_append_point_vec( &mut self, label: &'static [u8], points: &Vec<CompressedRistretto>, ) -> Result<(), ProofError>

fn validate_and_append_point( &mut self, label: &'static [u8], point: &CompressedRistretto, ) -> Result<(), ProofError>

Check that a point is not the identity, then append it to the transcript. Otherwise, return an error.

fn validate_and_append_cipher( &mut self, label: &'static [u8], point: &Ciphertext, ) -> Result<(), ProofError>

fn challenge_scalar(&mut self, label: &'static [u8]) -> Scalar

Compute a labeled challenge variable.